requirements.txt
fastapi==0.103.0
uvicorn==0.23.2
asyncpg==0.28.0
sqlalchemy==2.0.20
alembic==1.12.0
pydantic==2.4.2
pydantic-settings==2.0.3
httpx==0.24.1
python-jose[cryptography]==3.3.0
redis==4.5.0
structlog==23.1.0
fastapi-limiter==0.1.5
tenacity==8.2.3
pandas==2.0.3
reportlab==4.0.4
pytest==7.4.0
pytest-asyncio==0.21.1
